Add a method to render the plot to anything implementing Write
Plotly supported rendering the plot to HTML, and writing the output to a file, located at a user given path. Often however, it's desirable to split the process of rendering and the process of outputting the rendering. In addition, writing to a path is not always the most flexible, since you don't always have a file path. It can then be useful to be able to write the rendered entity to any writable buffer, which can be accomplished by using the std::io::Write trait instead of a path, which is what this change set does.