Add 'plotly_noembed' cargo feature
This adds a new `plotly_noembed` feature that can be used to reduce the
sizes of binaries by not embedding `plotly.min.js`.

Since askama can not handle compile-time flags, we had to duplicate the
`plot.html` and `static_plot.html` templates, where the new
`*_noembed.html` version only contains the CDN-variant.

The `use_local_plotly` option has been feature-gated such that it is
not available when using the new feature.

I verified that this works by building an application that depends on
plotly.rs (including optimization and LTO in all cases):

    no plotly:   4.96 MiB (5,201,920 bytes)
    master:      8.96 MiB (9,400,320 bytes)
    this branch: 5.44 MiB (5,701,632 bytes)

The difference between this branch and master is 3,698,688 bytes, which
is very close to the size of `plotly.min.js` (3,686,400 bytes). Just as
expected.