Update to askama 0.13.0
Rinja 0.2 was a fork of askama 0.12. Both projects were [merged again],
and new versions of the project will be released under the name askama.

The feature `"with-axum"` is not needed anymore. The implementation for
web-framework integrations was moved into its own crate. The `askama`
crate does not have any problems with feature flag incompatibilities
anymore.

[merged again]: <https://blog.guillaume-gomez.fr/articles/2025-03-19+Askama+and+Rinja+merge>